refactor: move normalization from ticket-list to filter
