+++ src/ticket-list/index.ts
@@ -1,0 +1,1 @@
+export { TicketsList } from './ticket-list';
