+++ src/ticket-list/model.ts
@@ -1,0 +1,5 @@
+import { $filteredTickets } from '../filter';
+
+export const $visibleTickets = $filteredTickets.map((tickets) =>
+  tickets.slice(0, 5),
+);
