--- src/tickets-list/index.ts
@@ -1,1 +1,0 @@
-export { TicketsList } from './tickets-list';
